// src/lib/api/generated/models/carStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type CarStatus = (typeof CarStatus)[keyof typeof CarStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarStatus = {
  active: 'active',
  sold: 'sold',
  offShelf: 'offShelf',
  pending: 'pending',
} as const