// src/lib/api/generated/models/sellerType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type SellerType = (typeof SellerType)[keyof typeof SellerType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SellerType = {
  individual: 'individual',
  dealer: 'dealer',
} as const