// src/lib/api/generated/models/sourcePlatformType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type SourcePlatformType = (typeof SourcePlatformType)[keyof typeof SourcePlatformType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SourcePlatformType = {
  xiaohongshu: 'xiaohongshu',
} as const