// src/lib/api/generated/models/transmissionType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type TransmissionType = (typeof TransmissionType)[keyof typeof TransmissionType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransmissionType = {
  automatic: 'automatic',
  manual: 'manual',
} as const